pyproject.toml
[tool.poetry]
name = "connect-api"
version = "0.1.0"
description = ""
authors = ["adhammhaydat <adhammohidat123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
django-cors-headers = "^3.10.1"
Pillow = "^8.4.0"
whitenoise = "^5.3.0"
pytz = "^2021.3"
fire = "^0.4.0"
requests = "^2.26.0"
psycopg2-binary = "^2.9.2"
django-environ = "^0.8.1"
gunicorn = "^20.1.0"
virtualenv = "^20.10.0"
django-nose = "^1.4.7"
coverage = "^6.2"

[tool.poetry.dev-dependencies]
Django = "^4.0"
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
